Moved all code to one file and created more functions
- Moved all of the code from separate Python scripts to one Python
script for initial development.
- Updated setup_database() function from database.py to include error
checking, and to give the option to use the function as a means to
opening the connection to the database as well.
- Created functions to handle errors in a simple way, prepare a Selenium
webdriver for the advanced search page, store the scraped parcel
numbers, and a simple shell function to return a fake parcel number.
- Lastly, I created a simple, visual integration test in the main
function to check database connection/insertion, along with connecting
the the auditor's website.