Began moving logic out of rough draft into separate files
- Now that I have delivered a preliminary CSV file to Prof McDonald, I
will now flesh out the logic and create a command-line application that
allows the user to choose many options, e.g. scraping data for one
parcel ID and then printing the results to the screen/save in a file,
pointing to a CSV file of parcel IDs, or scraping the parcel IDs
themselves.